Include headRepositoryId when creating a new PR
gh pr create does not currently handle creating pull requests across
repositories in the same organization. If a user attempts to create such a
PR, they will see a confusing error message:

  pull request create failed: GraphQL: Head sha can't be blank, Base sha can't be blank, No commits between innersource:main and sandbox:branch, Head ref must be a branch (createPullRequest)

According to the API documentation, to create such a pull request, the API
request must contain the headRepositoryId identifying where the PR branch
exists.

In the usual case where gh pr create is being called without arguments, the
head repository is already known. It is looked up in order to ensure the
local changes are pushed to the remote fork.

In submitPR, extract the repository ID from the headRepo. If the headRepo
is already a full repository object, just grab the ID from it. Otherwise
perform an additional API lookup to get the repository object and then grab
the ID from that.

Extend the CreatePullRequest API wrapper to pass the headRepositoryId if it
was given.

This partially fixes #10238, at least in cases where the head repository is
known. A different solution will be required for the general case where
--head is used.